@@ -1,8 +1,6 @@
 module stash.appscode.dev/ui-server
 
-go 1.23.0
-
-toolchain go1.24.4
+go 1.25.5
 
 require (
 	github.com/lnquy/cron v1.1.1
